+++ src/index.ts
@@ -1,5 +1,5 @@
 import { Lexer } from './lexer';
-import Parser from './parser';
+import Parser from './engine';
 import SymbolTable from './symboltable';
 
 module.exports = {
